+++ dfs.h
@@ -1,6 +1,6 @@
 #include <stdio.h>
 #include "stack.h"
 
-extern int dfsUpDown(int** array, int size);
+extern int dfsUpDown(short** array, int size);
 
-extern int dfsLeftRight(int** array, int size);
+extern int dfsLeftRight(short** array, int size);
